refactor, using parameter for sync request
