+++ argocd/orb.yml
@@ -23,34 +23,16 @@ commands:
       target:
         type: string
         description: "Commit hash to check for"
+      sync_request:
+        type: boolean
+        default: false
+        description: "whether to send a sync request before polling"
     steps:
       - run:
           name: "Wait for ArgoCD to sync"
           shell: /bin/bash -eo pipefail
           command: |
             include ./scripts/wait_for_sync.sh
-  sync:
-    description: "Sends a post request to the Argo api to sync the given app to its target commit hash. Then waits for sync to complete. Expects ARGOCD_TOKEN environment variable to be set."
-    parameters:
-      application:
-        type: string
-        description: "Application ID"
-      argocd_url:
-        type: string
-        description: "URL of ArgoCD server"
-      wait_for:
-        type: integer
-        default: 600
-        description: "Maximum time to wait for sync in seconds. Default 600."
-      target:
-        type: string
-        description: "Commit hash to check for"        
-    steps:
-      - run:
-          name: "sync request to app and wait for sync"
-          shell: /bin/bash -eo pipefail
-          command: |
-            include ./scripts/sync_and_wait.sh
 
 jobs:
   wait_for_sync:
@@ -69,31 +51,14 @@ jobs:
       target:
         type: string
         description: "Commit hash to check for"
+      sync_request:
+        type: boolean
+        default: false
+        description: "whether to send a sync request before polling"
     steps:
       - wait_for_sync:
           wait_for: << parameters.wait_for >>
           argocd_url: << parameters.argocd_url >>
           application: << parameters.application >>
           target: << parameters.target >>
-  sync:
-    executor: python
-    parameters:
-      wait_for:
-        type: integer
-        default: 600
-        description: "Maximum time to wait for in seconds. Default 600."    
-      application:
-        type: string
-        description: "Application ID"
-      argocd_url:
-        type: string
-        description: "URL of ArgoCD server"
-      target:
-        type: string
-        description: "Commit hash to check for"        
-    steps:
-      - sync:
-          wait_for: << parameters.wait_for >>
-          argocd_url: << parameters.argocd_url >>
-          application: << parameters.application >>
-          target: << parameters.target >>
+          sync_request: << parameters.sync_request >>
